Move feature specs to system specs
(We had both?)
